# api/openapi.yaml
openapi: 3.0.3
info:
  title: Go.Kitchen API (Go.Kitchen Service)
  description: |
    REST API сервиса доставки еды «Go.Kitchen».
    Предоставляет интерфейсы для покупателей (каталог заведений, меню, оформление заказов, отслеживание и отмена),
    а также для ресторанов-партнеров (управление меню, стоп-лист, остатки, очередь заказов, изменение статусов готовки и вебхуки).
  version: 1.0.0
  contact:
    name: Go.Kitchen Core Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Основной сервис Go.Kitchen (Core Kitchen Service)
  - url: http://localhost:8081
    description: Демо-сервис ресторана-партнера (Partner Restaurant Simulator)
...
paths:
...
  /api/v1/restaurants:
    get:
      operationId: listRestaurants
      tags:
        - Client API
      summary: Получить каталог доступных ресторанов
      description: |
        Возвращает список только активных заведений, готовых принимать заказы.
        Параметр `search` отвечает на вопрос «как человек находит нужное заведение»:
        регистронезависимый поиск подстроки по названию и описанию.
      parameters:
        - name: search
          in: query
          description: |
            Подстрока названия или описания заведения (регистронезависимо).
            Минимум 3 символа: более короткий паттерн не покрывается триграммным индексом
            и выродился бы в полное сканирование каталога, поэтому отклоняется с 400.
          schema:
            type: string
            minLength: 3
            maxLength: 100
          example: пицц
        - name: limit
          in: query
          description: Количество записей на странице. Значение вне диапазона отклоняется с 400.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Смещение выборки для пагинации
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список ресторанов успешно получен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/v1/restaurants/{restaurant_id}:
    get:
      operationId: getRestaurant
      tags:
        - Client API
      summary: Получить информацию о ресторане по ID
      parameters:
        - name: restaurant_id
          in: path
          required: true
          description: UUID заведения
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о заведении
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Restaurant'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/orders:
    get:
      operationId: listMyOrders
      tags:
        - Client API
      summary: История заказов покупателя
      description: |
        Аутентификация пользователей вне скоупа задания, поэтому покупатель называет себя сам
        в `user_id`, и параметр обязателен: без него ручка отдавала бы все заказы системы,
        а не историю одного человека. Запрос ложится на индекс `idx_orders_user_id`.
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: Фильтр по статусу заказа
          schema:
            type: string
            enum: [created, accepted, cooking, ready, delivering, completed, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Заказы покупателя, новые первыми
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'

    post:
      operationId: createOrder
      tags:
        - Client API
      summary: Оформить новый заказ
      description: |
        Атомарно проверяет активность заведения, блокирует позиции в БД, проверяет остатки,
        списывает их со склада, фиксирует моментальный снимок цен и отправляет вебхук партнеру.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Заказ успешно создан и передан в ресторан
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Недостаточно остатков на складе ресторана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: "OUT_OF_STOCK"
                  message: "one or more items are out of stock"
        '422':
          description: Ресторан неактивен, блюдо недоступно или блюда из разных заведений
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orders/{order_id}:
    get:
      operationId: getOrder
      tags:
        - Client API
      summary: Получить информацию о заказе (чек и статус)
      description: Карточка заказа содержит адрес доставки, комментарий и чек, поэтому покупатель называет себя параметром `user_id` — так же, как в истории заказов. Чужой заказ отвечает `404`, а не `403`, — по идентификатору нельзя проверить, что заказ существует.
      parameters:
        - name: order_id
          in: path
          required: true
          description: UUID заказа
          schema:
            type: string
            format: uuid
        - name: user_id
          in: query
          required: true
          description: UUID покупателя, оформившего заказ (заглушка вместо аутентификации)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали заказа и текущий статус
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/orders/{order_id}/cancel:
    post:
      operationId: cancelOrder
      tags:
        - Client API
      summary: Отменить заказ покупателем
      description: Покупатель может отменить заказ только до того, как кухня начала приготовление (`status = created`). При отмене остатки автоматически возвращаются на склад.
      parameters:
        - name: order_id
          in: path
          required: true
          description: UUID заказа
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          description: Заказ успешно отменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          description: Заказ уже передан на кухню или готовится и не может быть отменен покупателем
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    AdminBootstrapToken:
      type: apiKey
      in: header
      name: X-Bootstrap-Token
      description: |
        Токен площадки, разрешающий поставить новое заведение на учёт. Отдельный от партнёрского
        ключа: тот отвечает на вопрос «какое я заведение», этот — «можно ли добавить ещё одно».
        В стенде `docker compose` задан как `demo-bootstrap-token-000000000001`.
    PartnerApiKey:
      type: http
      scheme: bearer
      description: |
        Закрытый доступ для заведений. Ключ выдаётся один раз при регистрации
        (`POST /api/v1/partner/restaurants`) и определяет, от имени какого заведения
        выполняется запрос. Чужие категории, блюда и заказы отвечают `404`, чтобы по API
        нельзя было перебирать идентификаторы конкурентов.
        Альтернативно ключ принимается в заголовке `X-Api-Key`.
  schemas:
    Restaurant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Додо Пицца"
        description:
          type: string
          example: "Сеть пиццерий №1 в России"
        address:
          type: string
          example: "г. Москва, ул. Ленина, д. 10"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        menu_item_id:
          type: string
          format: uuid
        item_name:
          type: string
          example: "Чизбургер"
        price_cents:
          type: integer
          format: int64
          example: 25000
        quantity:
          type: integer
          example: 2
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, accepted, cooking, ready, delivering, completed, cancelled]
          example: "accepted"
        total_price_cents:
          type: integer
          format: int64
          example: 50000
        delivery_address:
          type: string
          example: "г. Москва, ул. Арбат, д. 12, кв. 44"
        comment:
          type: string
          example: "Пожалуйста, положите салфетки"
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status_history:
          type: array
          description: Аудит переходов статуса от старых к новым.
          items:
            $ref: '#/components/schemas/OrderStatusHistoryEntry'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderStatusHistoryEntry:
      type: object
      description: "Запись аудита: кто и когда перевёл заказ в новый статус."
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        from_status:
          type: string
          description: Пустая строка для самой первой записи (создание заказа).
          example: "created"
        to_status:
          type: string
          enum: [created, accepted, cooking, ready, delivering, completed, cancelled]
          example: "accepted"
        actor:
          type: string
          enum: [customer, restaurant, system]
          description: Инициатор перехода.
          example: "restaurant"
        comment:
          type: string
          example: "Заказ подтвержден рестораном и передан шеф-повару"
        created_at:
          type: string
          format: date-time
    CreateOrderItemRequest:
      type: object
      required:
        - menu_item_id
        - quantity
      properties:
        menu_item_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          example: 2
    CreateOrderRequest:
      type: object
      required:
        - user_id
        - restaurant_id
        - delivery_address
        - items
      properties:
        user_id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        delivery_address:
          type: string
          example: "г. Москва, ул. Ленина, д. 1"
        comment:
          type: string
          maxLength: 1000
          example: "Оставить у двери"
        items:
          type: array
          items:
            $ref: '#/components/schemas/CreateOrderItemRequest'
    CancelOrderRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 1000
          example: "Ошибся адресом доставки"
...
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "requested resource not found"
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
  responses:
...
    BadRequestError:
      description: Некорректный запрос или невалидный синтаксис данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "BAD_REQUEST"
              message: "invalid request data or syntax"
...
    NotFoundError:
      description: Запрашиваемый ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "requested resource not found"